Update weather struct and modify post loop
Updated the weather struct to have both celsius and fahrenheit
temperatures since Metabase won't allow me to convert data on the
fly. The sensor reads in celsius so it's just converting that value.

Switched the loop to be infinite now so we can start, and switch the
intervals to be 30mins now instead of 15seconds.